Move all exceptions to qwikstart.exceptions
