+++ qwikstart/operations/__init__.py
@@ -1,10 +1,9 @@
 from . import add_file, add_file_tree, insert_text, prompt_user
-from .base import BaseOperation, GenericOperation, OperationError
+from .base import BaseOperation, GenericOperation
 
 __all__ = [
     "BaseOperation",
     "GenericOperation",
-    "OperationError",
     "add_file",
     "add_file_tree",
     "insert_text",
