+++ qwikstart/parser/__init__.py
@@ -1,10 +1,4 @@
-from .core import OperationMapping, ParserError, get_operations_mapping
+from .core import OperationMapping, get_operations_mapping
 from .tasks import TaskDefinition, parse_task
 
-__all__ = [
-    "OperationMapping",
-    "ParserError",
-    "TaskDefinition",
-    "get_operations_mapping",
-    "parse_task",
-]
+__all__ = ["OperationMapping", "TaskDefinition", "get_operations_mapping", "parse_task"]
