+++ qwikstart/parser/operations.py
@@ -2,8 +2,9 @@ import collections
 from typing import Any, Dict, NamedTuple, Optional, Tuple, Type, Union
 
 from .. import utils
+from ..exceptions import TaskParserError
 from ..operations import GenericOperation
-from .core import OperationMapping, ParserError, get_operations_mapping
+from .core import OperationMapping, get_operations_mapping
 
 __all__ = ["OperationDefinition", "parse_operation"]
 
@@ -26,7 +27,7 @@ def parse_operation(
     op_def = normalize_op_definition(op_def)
 
     if op_def.name not in known_operations:
-        raise ParserError(f"Could not find operation named '{op_def.name}'")
+        raise TaskParserError(f"Could not find operation named '{op_def.name}'")
 
     operation_class = known_operations[op_def.name]
     return operation_class(**op_def.config)
@@ -42,7 +43,7 @@ def normalize_op_definition(
         return OperationDefinition(name=op_def, config={})
     elif isinstance(op_def, collections.abc.Mapping):
         if len(op_def) != 1:
-            raise ParserError(
+            raise TaskParserError(
                 "Operation definition with dict should only have a single key"
                 f", but given {op_def}"
             )
@@ -50,10 +51,10 @@ def normalize_op_definition(
         return OperationDefinition(name=op_name, config=op_def[op_name])
     elif isinstance(op_def, collections.abc.Sequence):
         if len(op_def) != 2:
-            raise ParserError(
+            raise TaskParserError(
                 "Operation definition with sequence should only have two items"
                 f", but given {op_def}"
             )
         return OperationDefinition(*op_def)
     else:
-        raise ParserError(f"Could not parse operation definition: {op_def}")
+        raise TaskParserError(f"Could not parse operation definition: {op_def}")
